requirements.txt
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
pillow>=10.0.0
joblib>=1.3.0
libsvm>=3.0.0
